+++ samples/kotlin/src/main/java/co/windly/limbosample/presentation/splash/SplashPresenter.kt
@@ -1,10 +1,10 @@
 package co.windly.limbosample.presentation.splash
 
 import co.windly.limbo.presenter.queue.LimboQueuePresenter
+import co.windly.limbo.utility.reactive.addTo
 import co.windly.limbo.utility.reactive.observeOnUi
 import co.windly.limbo.utility.reactive.subscribeOnComputation
 import io.reactivex.Observable
-import io.reactivex.rxkotlin.addTo
 import timber.log.Timber
 import java.util.concurrent.TimeUnit.MILLISECONDS
 import javax.inject.Inject